requirements.txt
python-dotenv
openai==1.17.1
notebook
notebook
pandas
pymongo
pygwalker
streamlit